Herschrijven code om Matching class uit te faseren en op te splitsen in match en check classes
